refactor(session): Move error codes and error types to separate file
- Extracted all error constant definitions and error types to synology_drive_error.go
- Fix: Updated error handling bug in Login() and Logout() methods
