+++ synology_drive_api/synology_drive_api.go
@@ -7,7 +7,6 @@ import (
 	"net/http"
 	"net/http/cookiejar"
 	"net/url"
-	"strconv"
 )
 
 type SynologySession struct {
@@ -24,75 +23,6 @@ func (s *SynologySession) sessionExpired() bool {
 	return s.sid == ""
 }
 
-// Common Synology API error codes (not exhaustive)
-//
-// These error codes are used in the Synology API documentation:
-// https://global.download.synology.com/download/Document/Software/DeveloperGuide/Os/DSM/All/enu/DSM_Login_Web_API_Guide_enu.pdf
-const SYNOLOGY_COMMON_ERROR_UNKNOWN = 100
-const SYNOLOGY_COMMON_ERROR_INVALID_PARAMETER = 101
-const SYNOLOGY_COMMON_ERROR_API_NOT_EXIST = 102
-const SYNOLOGY_COMMON_ERROR_METHOD_NOT_EXIST = 103
-const SYNOLOGY_COMMON_ERROR_VERSION_NOT_SUPPORTED = 104
-const SYNOLOGY_COMMON_ERROR_DOES_NOT_HAVE_PERMISSION = 105
-const SYNOLOGY_COMMON_ERROR_SESSION_TIMEOUT = 106
-const SYNOLOGY_COMMON_ERROR_SESSION_INTERRUPTED = 107
-const SYNOLOGY_COMMON_ERROR_UPLOAD_FILE_FAILED = 108
-const SYNOLOGY_COMMON_ERROR_NETWORK_UNSTABLE_OR_SYSTEM_BUSY = 109
-const SYNOLOGY_COMMON_ERROR_NETWORK_UNSTABLE_OR_SYSTEM_BUSY_2 = 110
-const SYNOLOGY_COMMON_ERROR_NETWORK_UNSTABLE_OR_SYSTEM_BUSY_3 = 111
-const SYNOLOGY_COMMON_ERROR_RESERVED_112 = 112
-const SYNOLOGY_COMMON_ERROR_RESERVED_113 = 113
-const SYNOLOGY_COMMON_ERROR_LOST_PARAMETER = 114
-const SYNOLOGY_COMMON_ERROR_UPLOAD_FILE_DISALLOWED = 115
-const SYNOLOGY_COMMON_ERROR_OPERATION_DISALLOWED_FOR_DEMO_SITE = 116
-const SYNOLOGY_COMMON_ERROR_NETWORK_UNSTABLE_OR_SYSTEM_BUSY_4 = 117
-const SYNOLOGY_COMMON_ERROR_NETWORK_UNSTABLE_OR_SYSTEM_BUSY_5 = 118
-const SYNOLOGY_COMMON_ERROR_INVALID_SESSION = 119
-const SYNOLOGY_COMMON_ERROR_RESERVED_120 = 120
-const SYNOLOGY_COMMON_ERROR_RESERVED_121 = 121
-const SYNOLOGY_COMMON_ERROR_RESERVED_122 = 122
-const SYNOLOGY_COMMON_ERROR_RESERVED_123 = 123
-const SYNOLOGY_COMMON_ERROR_RESERVED_124 = 124
-const SYNOLOGY_COMMON_ERROR_RESERVED_125 = 125
-const SYNOLOGY_COMMON_ERROR_RESERVED_126 = 126
-const SYNOLOGY_COMMON_ERROR_RESERVED_127 = 127
-const SYNOLOGY_COMMON_ERROR_RESERVED_128 = 128
-const SYNOLOGY_COMMON_ERROR_RESERVED_129 = 129
-const SYNOLOGY_COMMON_ERROR_RESERVED_130 = 130
-const SYNOLOGY_COMMON_ERROR_RESERVED_131 = 131
-const SYNOLOGY_COMMON_ERROR_RESERVED_132 = 132
-const SYNOLOGY_COMMON_ERROR_RESERVED_133 = 133
-const SYNOLOGY_COMMON_ERROR_RESERVED_134 = 134
-const SYNOLOGY_COMMON_ERROR_RESERVED_135 = 135
-const SYNOLOGY_COMMON_ERROR_RESERVED_136 = 136
-const SYNOLOGY_COMMON_ERROR_RESERVED_137 = 137
-const SYNOLOGY_COMMON_ERROR_RESERVED_138 = 138
-const SYNOLOGY_COMMON_ERROR_RESERVED_139 = 139
-const SYNOLOGY_COMMON_ERROR_RESERVED_140 = 140
-const SYNOLOGY_COMMON_ERROR_RESERVED_141 = 141
-const SYNOLOGY_COMMON_ERROR_RESERVED_142 = 142
-const SYNOLOGY_COMMON_ERROR_RESERVED_143 = 143
-const SYNOLOGY_COMMON_ERROR_RESERVED_144 = 144
-const SYNOLOGY_COMMON_ERROR_RESERVED_145 = 145
-const SYNOLOGY_COMMON_ERROR_RESERVED_146 = 146
-const SYNOLOGY_COMMON_ERROR_RESERVED_147 = 147
-const SYNOLOGY_COMMON_ERROR_RESERVED_148 = 148
-const SYNOLOGY_COMMON_ERROR_RESERVED_149 = 149
-const SYNOLOGY_COMMON_ERROR_SOURCE_IP_MISMATCH = 150
-
-// Synlogy Login API error codes
-const SYNOLOGY_LOGIN_ERROR_ACCOUNT_OR_PASSWORD_INCORRECT = 400
-const SYNOLOGY_LOGIN_ERROR_ACCOUNT_DISABLED = 401
-const SYNOLOGY_LOGIN_ERROR_PERMISSION_DENIED = 402
-const SYNOLOGY_LOGIN_ERROR_2FA_REQUIRED = 403
-const SYNOLOGY_LOGIN_ERROR_2FA_CODE_INCORRECT = 404
-const SYNOLOGY_LOGIN_ERROR_RESERVED_405 = 405
-const SYNOLOGY_LOGIN_ERROR_2FA_ENFORCED = 406
-const SYNOLOGY_LOGIN_ERROR_IP_BLOCKED = 407
-const SYNOLOGY_LOGIN_ERROR_CANNOT_CHANGE_EXPIRED_PASSWORD = 408
-const SYNOLOGY_LOGIN_ERROR_PASSWORD_EXPIRED = 409
-const SYNOLOGY_LOGIN_ERROR_PASSWORD_MUST_BE_CHANGED = 410
-
 // Synology Session name constant for API calls (private to this package)
 const synologySessionName = "SynologyDrive"
 
@@ -106,24 +36,6 @@ type synologyAuthResponseV3 struct {
 	Err struct {
 		Code int `json:"code"`
 	} `json:"error"`
-}
-
-type InvalidUrlError string
-
-func (e InvalidUrlError) Error() string {
-	return "invalid URL " + strconv.Quote(string(e)) + " in base_url"
-}
-
-type HttpError string
-
-func (e HttpError) Error() string {
-	return "http error " + strconv.Quote(string(e))
-}
-
-type SynologyError string
-
-func (e SynologyError) Error() string {
-	return "synology error " + strconv.Quote(string(e))
 }
 
 // NewSynologySession creates a new Synology API session with the provided credentials and base URL.
@@ -201,7 +113,7 @@ func (s *SynologySession) Login() error {
 
 	resp, err := s.httpGet(endpoint, params)
 	if err != nil {
-		return HttpError(err.Error())
+		return err
 	}
 
 	body, err := io.ReadAll(resp.Body)
@@ -242,7 +154,7 @@ func (s *SynologySession) Logout() error {
 
 	resp, err := s.httpGet(endpoint, params)
 	if err != nil {
-		return HttpError(err.Error())
+		return err
 	}
 
 	body, err := io.ReadAll(resp.Body)
